requirements.txt
Pillow
cryptography
tk